requirements.txt
Flask
pandas
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
geojson
xlrd
flask-cors